Refactoring: Move 'leveling.py' to agb/system
