+++ agb/system/message/message.py
@@ -15,7 +15,7 @@
 #    along with AlphaGameBot.  If not, see <https://www.gnu.org/licenses/>.
 
 import agb.cogwheel
-import agb.system.message.leveling
+import agb.system.leveling
 import agb.system.onboarding
 import agb.system.counting
 import mysql.connector
@@ -63,7 +63,7 @@ async def handleOnMessage(bot: commands.Bot,
     if tracking and CAN_USE_DATABASE:
         await agb.system.onboarding.initalizeNewUser(cnx, ctx.author.id, ctx.guild.id)
         await agb.system.counting.countPoints(ctx, cnx, CAN_USE_DATABASE, tracking)
-        await agb.system.message.leveling.handleMessageLevelUp(ctx, cnx, CAN_USE_DATABASE, tracking)
+        await agb.system.leveling.handleMessageLevelUp(ctx, cnx, CAN_USE_DATABASE, tracking)
     else:
         logging.debug("handleOnMessage: Not dispatching tracking functions because %s", 
                       "tracking is disabled" if not tracking else "the database is disabled")
